requirements.txt
opencv-python
numpy
face-recognition
pyaudio
wave
webrtcvad
whisper
pyttsx3
picamera2
sounddevice
wavio
SpeechRecognition
google-cloud-speech
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
simpleaudio
websocket-client
